Interface improvement
Replace any usage of a library sequence map with just a pure array of sequences. This cleans up the UI side and makes for one less distinct interface for the map.

Separate the spec for languages into a base interface and extensions for input and output formats.

Set up the actual 'adaptation' to contain a getter for the languages given injected dependencies (just utilities for building Svelte tooptlip components for now). What was previously passed around as the 'adaptation' is now 'PhoenixLanguages', which is I think more apt.

Also, fix the tooltips that broke when we extracated from aerie-ui.